translate operator types in list and edit
