+++ src/app/models/operator.model.ts
@@ -1,8 +1,18 @@
 import { JsonApiModel, JsonApiModelConfig, Attribute, BelongsTo, HasMany } from 'angular2-jsonapi';
 import { Fmu } from 'app/models/fmu.model';
 import { Country } from 'app/models/country.model';
-import { User } from 'app/models/user.model';
-import { OperatorTypes } from 'app/pages/fields/operators/operator-list.component';
+
+export enum OperatorTypes {
+  'Artisanal' = 'Artisanal',
+  'Community forest' = 'Community forest',
+  'Estate' = 'Estate',
+  'Industrial agriculture' = 'Industrial agriculture',
+  'Logging company' = 'Logging company',
+  'Mining company' = 'Mining company',
+  'Other' = 'Other',
+  'Sawmill' = 'Sawmill',
+  'Unknown' = 'Unknown'
+}
 
 @JsonApiModelConfig({
   type: 'operators'
